MGMT-8473: Feature support level indications in cluster details
